Add favicon assets and refactor server code
- Add various favicon and apple icon assets
- Add .prettierignore for Prettier configuration
- Update .editorconfig and .vscode/extensions.json
- Add Makefile for build automation
- Update README.md with new instructions
- Add browserconfig.xml and manifest.json for PWA support
- Refactor server code: add http.h, server.h, signal_handler.h, utils.h
- Add new source files: http.c, main.c, signal_handler.c, utils.c
- Remove src/simple_server.c
- Update index.html and obligation_register.code-workspace
- Add custom Modernizr build script
