+++ .vscode/extensions.json
@@ -1,8 +1,6 @@
 {
     "recommendations": [
-        "ms-vscode.cpptools",
         "ms-vscode.cpptools-extension-pack",
-        "ms-vscode.cpptools-themes",
         "dotenv.dotenv-vscode",
         "editorconfig.editorconfig",
         "github.copilot",
