+++ index.html
@@ -24,6 +24,25 @@
     <!-- Include Custom Styles -->
     <link rel="stylesheet" href="styles/reset.css" />
 
+    <!-- Favicons and Manifest -->
+    <link rel="apple-touch-icon" sizes="57x57" href="/apple-icon-57x57.png">
+    <link rel="apple-touch-icon" sizes="60x60" href="/apple-icon-60x60.png">
+    <link rel="apple-touch-icon" sizes="72x72" href="/apple-icon-72x72.png">
+    <link rel="apple-touch-icon" sizes="76x76" href="/apple-icon-76x76.png">
+    <link rel="apple-touch-icon" sizes="114x114" href="/apple-icon-114x114.png">
+    <link rel="apple-touch-icon" sizes="120x120" href="/apple-icon-120x120.png">
+    <link rel="apple-touch-icon" sizes="144x144" href="/apple-icon-144x144.png">
+    <link rel="apple-touch-icon" sizes="152x152" href="/apple-icon-152x152.png">
+    <link rel="apple-touch-icon" sizes="180x180" href="/apple-icon-180x180.png">
+    <link rel="icon" type="image/png" sizes="192x192"  href="/android-icon-192x192.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="96x96" href="/favicon-96x96.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png">
+    <link rel="manifest" href="/manifest.json">
+    <meta name="msapplication-TileColor" content="#ffffff">
+    <meta name="msapplication-TileImage" content="/ms-icon-144x144.png">
+    <meta name="theme-color" content="#ffffff">
+
     <style>
       .navbar-custom { background-color: #a3b383; }
       .navbar-icons i { font-size: 1.5rem; margin: 0 10px; background-color: #808d67; }
@@ -191,10 +210,7 @@
     ></script>
 
     <!-- Modernizr for HTML5 and CSS3 feature detection -->
-    <script
-      src="https://cdn.jsdelivr.net/npm/modernizr@3.13.1/lib/cli.min.js"
-      crossorigin="anonymous"
-    ></script>
+    <script src="scripts/modernizr-custom.js"></script>
 
     <!-- Polyfill.io for browser compatibility -->
     <script
