+++ src/signal_handler.c
@@ -1,0 +1,30 @@
+/**
+ * Copyright (C) 2024 Enveng Group - Adrian Gallo, Rohan Lonkar and Rhett Bachoup
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#include "../include/signal_handler.h"
+#include <stdlib.h>
+#include <unistd.h>
+
+extern int server_fd;
+
+/* Handle the SIGINT signal to gracefully shut down the server */
+void
+handle_signal (int signal)
+{
+  close (server_fd);
+  exit (EXIT_SUCCESS);
+}
